adds `version_message_with_relay` test
adds doc comment to `read_version`

renames codec/tests.rs to codec/tests/vectors.rs
